fix: address code review feedback on holo-tree API
1. commit_tree: drop required author/email params, use repo.author()
   and repo.committer() from git config/env with sensible fallback

2. SignatureRef time: the fragile string-formatted time hack in
   commit_tree is now only in the fallback path; the primary path
   uses gix's own signature resolution

3. resolve_tree_at_path: moved to holo-tree as create_tree_from_path()
   since it's a generic tree operation, not projection-specific

4. Re-export gix::ObjectId from holo-tree so consumers don't need
   gix as a direct dependency just for the type

5. Delete redundant holo-projector/Cargo.lock (workspace root is
   sufficient) and holo-projector/.gitignore (workspace target/ is
   at root)

6. Document write_child type-conflict behavior (errors if intermediate
   path component is a blob, unlike JS which silently replaces)
