+++ .gitignore
@@ -3,3 +3,5 @@ results/
 npm-debug.log
 .npmrc
 /target/
+holo-projector/target/
+holo-tree/target/
