--- holo-projector/.gitignore
@@ -1,1 +1,0 @@
-/target/
