+++ holo-tree/src/repo.rs
@@ -34,34 +34,87 @@ pub fn create_tree_from_ref(repo: &gix::Repository, git_ref: &str) -> Result<Mut
     Ok(MutableTree::new(tree_id))
 }
 
-/// Create a git commit object pointing to a tree.
+/// Create a MutableTree by navigating into a tree object at a subpath.
+///
+/// Given a tree OID and a slash-separated path, walks each component
+/// to find the final subtree. Returns an error if any component is missing.
+pub fn create_tree_from_path(
+    repo: &gix::Repository,
+    tree_id: ObjectId,
+    path: &str,
+) -> Result<MutableTree> {
+    use gix::bstr::ByteSlice;
+
+    if path == "." || path.is_empty() {
+        return Ok(MutableTree::new(tree_id));
+    }
+
+    let mut current = tree_id;
+    for component in path.split('/') {
+        if component.is_empty() || component == "." {
+            continue;
+        }
+
+        let obj = repo.find_object(current)?;
+        let tree = obj
+            .try_into_tree()
+            .map_err(|_| Error::NotATree(current.to_string()))?;
+
+        let entry = tree
+            .iter()
+            .filter_map(|e| e.ok())
+            .find(|e| {
+                e.filename()
+                    .to_str()
+                    .map(|s| s == component)
+                    .unwrap_or(false)
+            })
+            .ok_or_else(|| Error::PathNotFound {
+                component: component.to_string(),
+            })?;
+
+        current = entry.oid().to_owned();
+    }
+
+    Ok(MutableTree::new(current))
+}
+
+/// Create a git commit pointing to a tree.
+///
+/// Uses the repository's configured author/committer identity (from
+/// git config or `GIT_AUTHOR_NAME`/`GIT_COMMITTER_NAME` env vars).
+/// Falls back to "holo-tree" if no identity is configured.
 pub fn commit_tree(
     repo: &gix::Repository,
     tree_hash: ObjectId,
     parents: &[ObjectId],
     message: &str,
-    author_name: &str,
-    author_email: &str,
 ) -> Result<ObjectId> {
     use gix::objs::Commit;
 
-    let sig = gix::actor::SignatureRef {
-        name: author_name.into(),
-        email: author_email.into(),
-        time: &format!(
-            "{} +0000",
-            std::time::SystemTime::now()
-                .duration_since(std::time::UNIX_EPOCH)
-                .unwrap()
-                .as_secs()
-        ),
-    };
+    let author = repo
+        .author()
+        .and_then(|r| r.ok())
+        .map(|s| s.to_owned())
+        .transpose()
+        .ok()
+        .flatten()
+        .unwrap_or_else(default_signature);
+
+    let committer = repo
+        .committer()
+        .and_then(|r| r.ok())
+        .map(|s| s.to_owned())
+        .transpose()
+        .ok()
+        .flatten()
+        .unwrap_or_else(default_signature);
 
     let commit = Commit {
         tree: tree_hash,
         parents: parents.into(),
-        author: sig.to_owned().map_err(|e| Error::Git(e.to_string()))?,
-        committer: sig.to_owned().map_err(|e| Error::Git(e.to_string()))?,
+        author,
+        committer,
         encoding: None,
         message: message.into(),
         extra_headers: vec![],
@@ -87,4 +140,21 @@ pub fn update_ref(
     )
     .map_err(|e| Error::Git(e.to_string()))?;
     Ok(())
+}
+
+/// Fallback signature when git config has no author/committer.
+fn default_signature() -> gix::actor::Signature {
+    gix::actor::SignatureRef {
+        name: "holo-tree".into(),
+        email: "holo-tree@localhost".into(),
+        time: &format!(
+            "{} +0000",
+            std::time::SystemTime::now()
+                .duration_since(std::time::UNIX_EPOCH)
+                .unwrap()
+                .as_secs()
+        ),
+    }
+    .to_owned()
+    .expect("valid fallback signature")
 }
